+++ src/main.tsx
@@ -4,12 +4,15 @@ import { BrowserRouter } from 'react-router-dom';
 import './index.css';
 import App from './App.tsx';
 import { AuthProvider } from './context/AuthContext.tsx';
+import { ScaleToFit } from './components/shell/ScaleToFit.tsx';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <BrowserRouter>
       <AuthProvider>
-        <App />
+        <ScaleToFit>
+          <App />
+        </ScaleToFit>
       </AuthProvider>
     </BrowserRouter>
   </StrictMode>,